refactor e produto

+++ manager/src/main/java/com/silcom/manager/domain/service/ProdutoComplementoService.java
@@ -6,9 +6,11 @@ import java.util.List;
 import javax.transaction.Transactional;
 
 import com.silcom.manager.domain.exception.DuplicateKeyException;
+import com.silcom.manager.domain.exception.ResourceInUseException;
 import com.silcom.manager.domain.exception.ResourceNotFoundException;
 import com.silcom.manager.domain.model.ProdutoComplemento;
 import com.silcom.manager.domain.repository.ProdutoComplementoRepository;
+import com.silcom.manager.domain.repository.ProdutoRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -18,12 +20,16 @@ import org.springframework.stereotype.Service;
 @Service
 public class ProdutoComplementoService {
     
+    private static final String ID_IN_USE = "Produto cor id %d em uso";
     private static final String ALREADY_EXISTS = "Produto complemento nome '%s' ou sigla '%s' já existe";
     private static final String NOME_NOT_FOUND = "Não foram encontrados complementos de produto com o nome '%s'";
     private static final String ID_NOT_FOUND = "Produto complemento id %d não encontrado";
 
     @Autowired
     private ProdutoComplementoRepository produtoComplementoRepository;
+
+    @Autowired
+    private ProdutoRepository produtoRepository;
 
     public Page<ProdutoComplemento> findAll(Pageable pageable) {
         return produtoComplementoRepository.findAll(pageable);
@@ -48,7 +54,9 @@ public class ProdutoComplementoService {
 
     @Transactional
     public ProdutoComplemento insert(final ProdutoComplemento produtoComplemento) {
-        var produtoComplementoFormatado = this.formatObject(produtoComplemento);
+        var produtoComplementoFormatado = produtoComplemento;
+        produtoComplementoFormatado.format();
+        
         if (produtoComplementoRepository.existsByNomeIgnoreCase(produtoComplemento.getNome()) ||
             produtoComplementoRepository.existsBySiglaIgnoreCase(produtoComplemento.getSigla())) {
             throw new DuplicateKeyException(
@@ -59,12 +67,18 @@ public class ProdutoComplementoService {
 
     @Transactional
     public void delete(final Long id) {
+        if (produtoRepository.existsByTipoId(id)) {
+            throw new ResourceInUseException(
+                String.format(ID_IN_USE, id));
+        }
         produtoComplementoRepository.delete(this.findById(id));
     }
 
     @Transactional
     public ProdutoComplemento update(final Long id, final ProdutoComplemento produtoComplemento) {
-        var produtoComplementoFormatado = this.formatObject(produtoComplemento);
+        var produtoComplementoFormatado = produtoComplemento;
+        produtoComplementoFormatado.format();
+
         var produtoComplementoRecovered = this.findById(id);
         if ((!produtoComplementoRecovered.getNome().equals(produtoComplementoFormatado.getNome()) &&
             produtoComplementoRepository.existsByNomeIgnoreCase(produtoComplemento.getNome()) ) ||
@@ -77,11 +91,6 @@ public class ProdutoComplementoService {
         produtoComplementoFormatado.setDataCriacao(produtoComplementoRecovered.getDataCriacao());
         produtoComplementoFormatado.setDataAtualizacao(OffsetDateTime.now());
         return produtoComplementoRepository.save(produtoComplementoFormatado);
-    }
-
-    private ProdutoComplemento formatObject(final ProdutoComplemento produtoComplemento) {
-        produtoComplemento.setSigla(produtoComplemento.getSigla().toUpperCase());
-        return produtoComplemento;
     }
 
 }
