+++ manager/src/main/java/com/silcom/manager/domain/service/ProdutoTipoService.java
@@ -6,8 +6,10 @@ import java.util.List;
 import javax.transaction.Transactional;
 
 import com.silcom.manager.domain.exception.DuplicateKeyException;
+import com.silcom.manager.domain.exception.ResourceInUseException;
 import com.silcom.manager.domain.exception.ResourceNotFoundException;
 import com.silcom.manager.domain.model.ProdutoTipo;
+import com.silcom.manager.domain.repository.ProdutoRepository;
 import com.silcom.manager.domain.repository.ProdutoTipoRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,12 +20,16 @@ import org.springframework.stereotype.Service;
 @Service
 public class ProdutoTipoService {
     
+    private static final String ID_IN_USE = "Produto tipo id %d em uso";
     private static final String ALREADY_EXISTS = "Produto tipo nome '%s' ou sigla '%s' já existe";
     private static final String NOME_NOT_FOUND = "Não foram encontrados cpres de produto com o nome '%s'";
     private static final String ID_NOT_FOUND = "Produto tipo id %d não encontrado";
 
     @Autowired
     private ProdutoTipoRepository produtoTipoRepository;
+
+    @Autowired
+    private ProdutoRepository produtoRepository;
 
     public Page<ProdutoTipo> findAll(Pageable pageable) {
         return produtoTipoRepository.findAll(pageable);
@@ -48,7 +54,9 @@ public class ProdutoTipoService {
 
     @Transactional
     public ProdutoTipo insert(final ProdutoTipo produtoTipo) {
-        var produtoTipoFormatado = this.formatObject(produtoTipo);
+        var produtoTipoFormatado = produtoTipo;
+        produtoTipoFormatado.format();
+
         if (produtoTipoRepository.existsByNomeIgnoreCase(produtoTipo.getNome()) ||
             produtoTipoRepository.existsBySiglaIgnoreCase(produtoTipo.getSigla())) {
             throw new DuplicateKeyException(
@@ -59,12 +67,18 @@ public class ProdutoTipoService {
 
     @Transactional
     public void delete(final Long id) {
+        if (produtoRepository.existsByTipoId(id)) {
+            throw new ResourceInUseException(
+                String.format(ID_IN_USE, id));
+        }
         produtoTipoRepository.delete(this.findById(id));
     }
 
     @Transactional
     public ProdutoTipo update(final Long id, final ProdutoTipo produtoTipo) {
-        var produtoTipoFormatado = this.formatObject(produtoTipo);
+        var produtoTipoFormatado = produtoTipo;
+        produtoTipoFormatado.format();
+
         var produtoTipoRecovered = this.findById(id);
         if ((!produtoTipoRecovered.getNome().equals(produtoTipoFormatado.getNome()) &&
             produtoTipoRepository.existsByNomeIgnoreCase(produtoTipo.getNome()) ) ||
@@ -77,11 +91,6 @@ public class ProdutoTipoService {
         produtoTipoFormatado.setDataCriacao(produtoTipoRecovered.getDataCriacao());
         produtoTipoFormatado.setDataAtualizacao(OffsetDateTime.now());
         return produtoTipoRepository.save(produtoTipoFormatado);
-    }
-
-    private ProdutoTipo formatObject(final ProdutoTipo produtoTipo) {
-        produtoTipo.setSigla(produtoTipo.getSigla().toUpperCase());
-        return produtoTipo;
     }
 
 }
